Tune down verbosity for suspected retransmits
There are cases when control packet is not acked fast enought,
is not sent fast enough - for example when handling PUSH_REPLY,
which requires setting up tunnel. In those cases control packet
will be retransmitted.

OpenVPN 2 changes packet-id on retransmission, so it passes
replay protection and got rejected as a replay by reliability layer, which
checks another packet-id (sequence id) which is used to assemble
our TCP-like stream.

OpenVPN 3, however, doesn't change packet-id on retransmission,
which triggers replay protection and causes level 1 nonfatal errors
in logs.

When replay protection sees the packet with the same timestamp
and packet-id and previously received one, this is likely retransmission
from OpenVPN 3. To not to scare users, tune verbosity down in this case.