Fix handling of inactivity [bytes] parameter
When [bytes] value exceeded INT_MAX, it was handled
incorrectly. On Linux and MinGW, it was treated as 0, effectively
disabling the feature. Starting from 2.5.3 we switched to MSVC,
which treated large value as INT_MAX. This caused unexpected
disconnects for setups which were working due to this bug.

Use uint64 to represent inactivity [bytes] and warn if value
exceeds INT_MAX to notify users about possible behavior change.

Signed-off-by: Lev Stipakov <lev@openvpn.net>