Fix M_ERRNO behavior on Windows
We use M_ERRNO flag in logging to display error code
and error message. This has been broken on Windows,
where we use error code from GetLastError() and
error description from strerror(). strerror() expects
C runtime error code, which is quite different from
last error code from WinAPI call. As a result, we got
incorrect error description.

The ultimate fix would be introducing another flag
for WinAPI errors, like M_WINERR and use either that or
M_ERRNO depends on context. However, the change would be
quite intrusive and in some cases it is hard to say which
one to use without looking into internals.

Instead we stick to M_ERRNO and in Windows case we
first try to obtain error code from GetLastError() and
if it returns ERROR_SUCCESS (which is 0), we assume that
we have C runtime error and use errno. To get error
description we use strerror_win32() with GetLastError()
and strerror() with errno.

strerror_win32() uses FormatMessage() internally, which
is the right way to get WinAPI error description.

While on it, add a new flag M_SKERR to display socket
error. On Linux it is resolved into M_ERRNO and on Windows
to an conjunction of another new flag M_WSAERR and M_ERRNO.
With M_WSAERR we use WSAGetLastError() to get error code
and strerror_win32() to get error description.
WSAGetLastError() is a correct way to get WinSock error code,
according to documentation.

Signed-off-by: Lev Stipakov <lev@openvpn.net>