vcpkg-ports: remove openssl port
We've added functionality to openvpn-build

  f83f21ef "windows-msi: add ability to specify port overlays in
  openvpn-build"

(which is used for release building) to use overlay ports, which
enables us to build releases with updated ports (like openssl)
without making changes to openvpn repo. This means that openssl port
overlay from this repo could be removed.

Developers can still use openvpn repo for Windows development and
use whatever openssl version is provided by vcpkg official repo. When
we need to build a release and the latest openssl is not there, we add
port overlay to openvpn-build.

While on it, bump vcpkg commit id in github actions.

Signed-off-by: Lev Stipakov <lev@openvpn.net>