vcpkg: use the latest versions of dependency ports
By removing builin-baseline constraint, we force
vcpkg to use latest versions of ports. As a side effect,
this bumps OpenSSL to 3.0.7.

While on it, update vcpkg commit in GHA.

We also have to apply a workaround for GHA
to make it use our vcpkg checkout instead of c:\vcpkg,
see

 https://github.com/actions/runner-images/issues/6376
 https://github.com/lukka/run-vcpkg/issues/170

Signed-off-by: Lev Stipakov <lev@openvpn.net>