msvc: always regenerate config-msvc.h
There is no easy way to detect whether this information
is outdated. One possible optimisation would be in the
Python code to actually only rewrite the file if it has
changed. Although VC is already pretty good in detecting
whether the code has actually changed so this might not
be worth the additional complexity.

While here, change some leading whitespace to tabs as
expected in Makefile.

Signed-off-by: Frank Lichtenheld <frank@lichtenheld.com>