tun: move print_windows_driver() out of tun.h
We got warnings from MinGW about function being defined
but not used when compiling modules which include tun.h.

This function is not defined as inline, so its definition
should not be in header. Since this is not a performance
critical, no need to make it inline.

Leave declaration in tun.h and move definition to tun.c.

Signed-off-by: Lev Stipakov <lev@openvpn.net>