vcpkg-ports/pkcs11-helper: rename patches to make file names shorter
Remove the number prefixes we do not use anyway.

And then rename Allow-the-build-to-succeed-if-configured-with-disabl.patch
to make it short enough that it can fit in a tar archive
built with --old-archive. Otherwise make dist might not include it:

tar: openvpn-2.7_git/contrib/vcpkg-ports/pkcs11-helper/Allow-the-build-to-succeed-if-configured-with-disabl.patch:
file name is too long (max 99); not dumped

Change-Id: I5155a09972124c8da03a7618693b057e9e3c3302
Signed-off-by: Frank Lichtenheld <frank@lichtenheld.com>